Convert to a valid nimble project
Now you can build this locally using:

    nimble build
    (not need to do even "nim c src/hn_reader.nim")

Or install it locally using:

    nimble install

Anyone on the world wide web can install this using:

    nimble install https://github.com/massung/hn-reader

---

One difference is that now the binary/executable is "hn_reader"
instead of "reader" (on unix systems) .. the same with .exe extension
on Windows systems.

---

PS: It's canonical to not have hyphens in Nim project names.. may I
suggest that you rename this repo from hn-reader to hn_reader?